requirements.txt
pyTelegramBotAPI==4.12.0
beautifulsoup4
python-dotenv
feedparser